sched: Optimize scheduler_ipi()
The current scheduler_ipi() function has 2 special conditions to
check, and it occured to me it's possible collapse it into a single
test.

This avoids a bunch of loads (got_nohz_idle_kick()) from the
fast-path and prepares infrastructure to avoid further condition
checks.

The trick is to enqueue a 'special' task to be woken, which then will
ensure the wake_list is not empty and we'll continue to run the
special code.

We use rq->idle as special task, since the idle task cannot sleep and
thus will never have to be woken. We do however need to make sure we
cannot enqueue it more than once.

XXX: write comments with the barriers
XXX: introduces 'harmless':
     NOHZ: local_softirq_pending 82
     probably due to now calling raise_softirq*() from new places,
     should be fixable.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>