sched: Add VCPU aware SMT scheduling
Mitigate the EPT-L1TF SMT problem by ensuring a VCPU thread only ever
shares the core with another VCPU of the _same_ VM or idle.

This is accomplished with 4 hooks (2 in schedule(), 1 in the IPI
handler and 1 in pick_next_task_fair()) all guarded by a static key.

The pick_next_task_fair() hook tries to, given some fairness
constraints, pick a task matching the current core filter. The post
schedule() hook checks if the selected task matches the filter, if not
it cancels the current filter (if any) and tries to establish a new
filter where appropriate. The pre schedule() hook is used to note
completion of the previous task, both for establishing filters and
tracking filter usage.

The scheme doesn't sync at VMENTER/VMEXIT but at (VCPU) task switch
points and as such 'leaks' host kernel state but not other guest/user
state.

In particular host scheduler and kvm state is not deemed sensitive.

A further assumption is that sending IPIs to SMT siblings is
sufficient to kick them out of guest mode. Given they all share the
same core and raising an IPI is an expensive and 'serializing'
operation this is not an unreasonable assumption and it removes a
number of icky wait-for points.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>