sched: Optimize VCPU trips into userspace
VCPU threads frequently drop out of guest mode and require user-space
emulation for various things. Instead of cancelling the core filter
and re-establishing it, implement a PAUSE state that kicks the other
SMT siblings into the scheduler and (busy) waits for the emulation to
be completed.

If either a waiting sibling needs to reschedule or the emulation
thread gets preempted, we cancel the core filter and let things
re-establish themselves when we resume the VCPU.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>